Refactor assetPOST API and useForm hook in uploadAssetUI.jsx
